+++ CHANGELOG.md
@@ -7,6 +7,8 @@ item (Added, Changed, Depreciated, Removed, Fixed, Security).
 ## [2026.03]
 
 - Updated: Don't adjust left/right padding on Group blocks with color themes.
+- Updated: Refactor Horizontal Tabs block into a dynamic block. Add reordering functionality. 
+- Updated: Refactor Vertical Tabs block to dynamic block; allow sorting.
 
 ## [2026.02]